net: mvpp2: Add Group-of-Ports and Net-Compex configuration for PPv2.2
This patch adds the GoP (Group of Ports) and NetC (Net Complex) setup to
the Marvell mvpp2 ethernet driver. This code is mostly copied from the
Marvell U-Boot version and was written by Stefan Chulski.

Currently only MAC2 (SGMII) and MAC3 (RGMII) are tested on the
Armada-7040 DB. Testing of the other MACs on other boards in all possible
PHY interface modes still needs to be done.

Signed-off-by: Stefan Roese <sr@denx.de>
Cc: Stefan Chulski <stefanc@marvell.com>
Cc: Kostya Porotchkin <kostap@marvell.com>
Cc: Nadav Haklai <nadavh@marvell.com>