mtd: nand: pxa3xx_nand: Increase initial buffer size
The initial buffer is used for the initial commands used to detect
a flash device (STATUS, READID and PARAM).

ONFI param page is 256 bytes, and there are three redundant copies
to be read. JEDEC param page is 512 bytes, and there are also three
redundant copies to be read. Hence this buffer should be at least
512 x 3. This commits rounds the buffer size to 2048.

[ Linux commit c16340973fcb6461474a9f811f7f3ff2f946b24c ]

Cc: Ezequiel Garcia <ezequiel at vanguardiasur.com.ar>
Signed-off-by: Chris Packham <judge.packham at gmail.com>
Reviewed-by: Ofer Heifetz <oferh@marvell.com>
Tested-by: Ofer Heifetz <oferh@marvell.com>